# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.saveCanvasRequest:
    properties:
      document:
        description: 画布文档内容，必须是合法的 JSON 对象
        type: object
      title:
        description: 画布标题
        example: 我的画布
        maxLength: 128
        minLength: 1
        type: string
    required:
    - document
    - title
    type: object
  controller.submitCanvasRequest:
    properties:
      access_level:
        description: 访问级别：1 公开，2 仅发布者 VIP 可见
        enum:
        - 1
        - 2
        example: 1
        type: integer
      required_plan_level:
        description: access_level 为 2 时要求的 VIP 等级
        maximum: 10
        minimum: 1
        type: integer
    required:
    - access_level
    type: object
  model.Canvas:
    properties:
      access_level:
        description: 访问级别：0 私有，1 公开，2 仅发布者 VIP 可见
        enum:
        - 0
        - 1
        - 2
        example: 0
        type: integer
      created_at:
        description: 创建时间
        type: string
      document:
        description: 画布文档内容
        type: object
      id:
        description: 画布 ID
        example: 1
        type: integer
      publish_status:
        description: 发布状态：0 草稿，1 待审核，2 已发布，3 已驳回
        enum:
        - 0
        - 1
        - 2
        - 3
        example: 0
        type: integer
      reject_reason:
        description: 审核驳回原因
        type: string
      required_plan_level:
        description: 要求的发布者 VIP 等级
        maximum: 10
        minimum: 1
        type: integer
      reviewed_at:
        description: 审核时间
        type: string
      reviewed_by:
        description: 审核人用户 ID
        type: integer
      submitted_at:
        description: 提交审核时间
        type: string
      title:
        description: 画布标题
        example: 我的画布
        type: string
      updated_at:
        description: 最后更新时间
        type: string
      user_id:
        description: 所属用户 ID
        example: 10
        type: integer
    type: object
  model.CanvasSummary:
    properties:
      access_level:
        description: 访问级别：0 私有，1 公开，2 仅发布者 VIP 可见
        enum:
        - 0
        - 1
        - 2
        example: 0
        type: integer
      created_at:
        description: 创建时间
        example: "2026-09-10T10:00:00+08:00"
        type: string
      id:
        description: 画布 ID
        example: 1
        type: integer
      publish_status:
        description: 发布状态：0 草稿，1 待审核，2 已发布，3 已驳回
        enum:
        - 0
        - 1
        - 2
        - 3
        example: 0
        type: integer
      title:
        description: 画布标题
        example: 我的画布
        type: string
      updated_at:
        description: 最后更新时间
        example: "2026-09-10T12:00:00+08:00"
        type: string
    type: object
  resp.ErrorInfo:
    properties:
      code:
        description: 便于程序识别的错误码
        example: INVALID_PARAM
        type: string
      message:
        description: 便于阅读的错误信息
        example: 参数无效
        type: string
    type: object
  resp.Pagination:
    properties:
      page:
        description: 当前页码
        example: 1
        type: integer
      size:
        description: 每页记录数
        example: 20
        type: integer
      total:
        description: 符合条件的记录总数
        example: 42
        type: integer
    type: object
  resp.Response:
    properties:
      data:
        description: 响应数据
      error:
        allOf:
        - $ref: '#/definitions/resp.ErrorInfo'
        description: 请求失败时的错误信息
      pagination:
        allOf:
        - $ref: '#/definitions/resp.Pagination'
        description: 列表接口的分页信息
      success:
        description: 请求是否成功
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Go Web HTTP 接口文档。
  title: Go Web API
  version: "1.0"
paths:
  /canvases:
    get:
      parameters:
      - description: 页码，从 1 开始
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页记录数
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CanvasSummary'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取画布列表
      tags:
      - canvases
    post:
      consumes:
      - application/json
      parameters:
      - description: 画布内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.saveCanvasRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Canvas'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建画布
      tags:
      - canvases
  /canvases/{id}:
    get:
      parameters:
      - description: 画布 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Canvas'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取画布详情
      tags:
      - canvases
  /canvases/{id}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 画布 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 发布设置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.submitCanvasRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Canvas'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 提交画布审核
      tags:
      - canvases
  /canvases/deleteById/{id}:
    post:
      parameters:
      - description: 画布 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除画布
      tags:
      - canvases
  /canvases/updateById/{id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 画布 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 画布内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.saveCanvasRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Canvas'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新画布
      tags:
      - canvases
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer {token}"，例如：Bearer eyJhbGciOiJIUzI1Ni...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"